# Optimize decode performance. Fixes #30.