With careful coaxing of the compiler to eliminate some bounds checks, we can get a nice speedup:

```
 name                    control ns/iter         branch ns/iter          diff ns/iter   diff % 
 decode_100b             112 (892 MB/s)          85 (1176 MB/s)                   -27  -24.11% 
 decode_100b_reuse_buf   92 (1086 MB/s)          64 (1562 MB/s)                   -28  -30.43% 
 decode_10mib            9,703,318 (1080 MB/s)   6,297,411 (1665 MB/s)     -3,405,907  -35.10% 
 decode_10mib_reuse_buf  7,592,036 (1381 MB/s)   4,671,137 (2244 MB/s)     -2,920,899  -38.47% 
 decode_30mib            28,165,896 (1116 MB/s)  19,467,132 (1615 MB/s)    -8,698,764  -30.88% 
 decode_30mib_reuse_buf  23,294,619 (1350 MB/s)  14,645,743 (2147 MB/s)    -8,648,876  -37.13% 
 decode_3b               45 (88 MB/s)            45 (88 MB/s)                       0    0.00% 
 decode_3b_reuse_buf     23 (173 MB/s)           25 (160 MB/s)                      2    8.70% 
 decode_3kib             2,165 (1418 MB/s)       1,325 (2318 MB/s)               -840  -38.80% 
 decode_3kib_reuse_buf   2,127 (1444 MB/s)       1,291 (2379 MB/s)               -836  -39.30% 
 decode_3mib             2,685,599 (1171 MB/s)   1,820,355 (1728 MB/s)       -865,244  -32.22% 
 decode_3mib_reuse_buf   2,220,699 (1416 MB/s)   1,354,520 (2322 MB/s)       -866,179  -39.00% 
 decode_500b             380 (1315 MB/s)         246 (2032 MB/s)                 -134  -35.26% 
 decode_500b_reuse_buf   380 (1315 MB/s)         228 (2192 MB/s)                 -152  -40.00% 
 decode_50b              72 (722 MB/s)           63 (825 MB/s)                     -9  -12.50% 
 decode_50b_reuse_buf    54 (962 MB/s)           43 (1209 MB/s)                   -11  -20.37% 
```

Now decoding is faster than encoding.

Also, bad padding errors are more consistent: they always refer to the _first_ padding character when a string of padding characters is detected as invalid for some reason (too long, followed by non-padding, etc).